Create brier loss
The Brier score, named for Glenn Brier, calculates the mean squared error between predicted
probabilities and the expected values. The score summarizes the magnitude of the error
in the probability forecasts. The error score is always between 0.0 and 1.0, where a model
with perfect skill has a score of 0.0. The Brier score can be calculated in Python using the
brier score loss() function in scikit-learn